+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/AggregatabilityValidationSpec.groovy
@@ -73,7 +73,7 @@ class AggregatabilityValidationSpec extends Specification {
         tg.getDimensions() >> dimensionDict.apiNameToDimension.values()
         table = new LogicalTable("name", DAY, tg)
         dimensionDict.apiNameToDimension.values().each {
-            DimensionColumn.addNewDimensionColumn(table, it, new PhysicalTable("abc", DAY.buildZonedTimeGrain(DateTimeZone.UTC), [:]))
+            DimensionColumn.addNewDimensionColumn(table, it)
         }
     }