+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/ThetaSketchIntersectionReportingResources.groovy
@@ -118,7 +118,7 @@ class ThetaSketchIntersectionReportingResources extends Specification {
         table = new LogicalTable("NETWORK", DAY, tableGroup)
 
         for (Dimension dim : tableGroup.getDimensions()) {
-            DimensionColumn.addNewDimensionColumn(table, dim, physicalTable)
+            DimensionColumn.addNewDimensionColumn(table, dim)
         }
 
         metricDict = new MetricDictionary()